revamp ndf
