+++ ndmanager/API/iaea/__init__.py
@@ -1,0 +1,3 @@
+from ndmanager.API.iaea.sublibrary import IAEASublibrary
+from ndmanager.API.iaea.library import IAEALibrary
+from ndmanager.API.iaea.iaea import IAEA
