+++ ndmanager/CLI/fetcher/listlibs.py
@@ -3,35 +3,43 @@
 import argparse as ap
 import textwrap
 
-from ndmanager.data import ENDF6_LIBS, NDMANAGER_ENDF6
+from ndmanager.API.iaea import IAEA
+from ndmanager.data import NDMANAGER_ENDF6
 from ndmanager.format import footer, get_terminal_size, header
 
-
-def listlibs_parser(subparsers: ap._SubParsersAction):
-    """Add the parser for the 'ndf list' command to a subparser object
-
-    Args:
-        subparsers (argparse._SubParsersAction): An argparse subparser object
-    """
-    parser = subparsers.add_parser(
-        "list", help="List libraries compatible with NDManager"
-    )
-    parser.set_defaults(func=listlibs)
-
-
-def listlibs(_args):
-    """List the libaries available for download with NDManager."""
-    col, _ = get_terminal_size()
-    lst = []
-    lst.append(header("Available libraries"))
-    for libname, libdict in ENDF6_LIBS.items():
-        fancyname = libdict["fancyname"]
-        if (NDMANAGER_ENDF6 / libname).exists():
-            check = "✓"
+class NdfListCommand:
+    def __init__(self, args: ap.Namespace) -> None:
+        self.args = args
+        if not IAEA.is_cached():
+            print("Initializing IAEA database...")
+            self.iaea = IAEA()
         else:
-            check = " "
-        s = f"{libname:<8} {fancyname:<15} [{check}]: {libdict['info']}"
-        s = textwrap.wrap(s, initial_indent="", subsequent_indent=30 * " ", width=col)
-        lst.append("\n".join(s))
-    lst.append(footer())
-    print("\n".join(lst))
+            self.iaea = IAEA()
+
+        col, _ = get_terminal_size()
+        lst = []
+        lst.append(header("Available libraries"))
+        for libname in  self.iaea.aliases:
+            libdata = self.iaea[libname]
+            fancyname = libdata.name.rstrip("/")
+            if (NDMANAGER_ENDF6 / libname).exists():
+                check = "✓"
+            else:
+                check = " "
+            s = f"{libname:<10} {fancyname:<15} [{check}]: {libdata.library}"
+            s = textwrap.wrap(s, initial_indent="", subsequent_indent=30 * " ", width=col)
+            lst.append("\n".join(s))
+        lst.append(footer())
+        print("\n".join(lst))
+
+    @classmethod
+    def parser(cls, subparsers):
+        """Add the parser for the 'ndf list' command to a subparser object
+
+        Args:
+            subparsers (argparse._SubParsersAction): An argparse subparser object
+        """
+        parser = subparsers.add_parser(
+            "list", help="List libraries compatible with NDManager"
+        )
+        parser.set_defaults(func=cls)
