+++ ndmanager/CLI/fetcher/remove.py
@@ -5,33 +5,29 @@ import shutil
 
 from ndmanager.data import NDMANAGER_ENDF6
 
+class NdfRemoveCommand:
+    def __init__(self, args: ap.Namespace) -> None:
+        libraries = [NDMANAGER_ENDF6 / lib for lib in args.library]
+        for library in libraries:
+            if library.exists():
+                shutil.rmtree(library)
 
-def remove_parser(subparsers: ap._SubParsersAction):
-    """Add the parser for the 'ndf info' command to a subparser object
+    @classmethod
+    def parser(cls, subparsers):
+        """Add the parser for the 'ndf info' command to a subparser object
 
-    Args:
-        subparsers (argparse._SubParsersAction): An argparse subparser object
-    """
-    parser = subparsers.add_parser(
-        "remove", help="Remove one or more installed ENDF6 libraries"
-    )
-    parser.add_argument(
-        "library",
-        type=str,
-        help="Names of the libraries to remove",
-        action="extend",
-        nargs="+",
-    )
-    parser.set_defaults(func=remove)
+        Args:
+            subparsers (argparse._SubParsersAction): An argparse subparser object
+        """
+        parser = subparsers.add_parser(
+            "remove", help="Remove one or more installed ENDF6 libraries"
+        )
+        parser.add_argument(
+            "library",
+            type=str,
+            help="Names of the libraries to remove",
+            action="extend",
+            nargs="+",
+        )
+        parser.set_defaults(func=cls)
 
-
-def remove(args: ap.Namespace):
-    """Uninstall a library from the NDManager database
-
-    Args:
-        args (ap.Namespace): The argparse object containing the command line argument
-    """
-    libraries = [NDMANAGER_ENDF6 / lib for lib in args.library]
-    for library in libraries:
-        if library.exists():
-            shutil.rmtree(library)
