+++ ndmanager/__init__.py
@@ -2,8 +2,8 @@
 
 __version__="0.3.3dev"
 
+from .API.iaea import IAEA
 from .API.endf6 import Endf6
-from .API.iaea import download_endf6, fetch_sublibrary_list
 from .API.nuclide import Nuclide
 from .API.sha1 import compute_file_sha1
 from .API.utils import get_endf6
